Reintroduce extension split
