+++ Sources/iOS/Extensions/ListAdapter+UITableViewDataSource.swift
@@ -1,0 +1,55 @@
+import UIKit
+
+/// MARK: - UITableViewDataSource
+extension ListAdapter: UITableViewDataSource {
+
+  /**
+   Tells the data source to return the number of rows in a given section of a table view. (required)
+
+   - Parameter tableView	: The table-view object requesting this information.
+   - Parameter section: An index number identifying a section in tableView.
+   - Returns: The number of rows in section.
+   */
+  public func tableView(tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
+    return spot.component.items.count
+  }
+
+  /**
+   Asks the data source for a cell to insert in a particular location of the table view. (required)
+
+   - Parameter tableView	: A table-view object requesting the cell.
+   indexPath
+   - Parameter indexPath: An index path locating a row in tableView.
+   - Returns: An object inheriting from UITableViewCell that the table view can use for the specified row. Will return the default table view cell for the current component based of kind.
+   */
+  public func tableView(tableView: UITableView, cellForRowAtIndexPath indexPath: NSIndexPath) -> UITableViewCell {
+    if indexPath.item < spot.component.items.count {
+      spot.component.items[indexPath.item].index = indexPath.row
+    }
+
+    let reuseIdentifier = spot.identifier(indexPath)
+    let cell: UITableViewCell = tableView
+      .dequeueReusableCellWithIdentifier(reuseIdentifier, forIndexPath: indexPath)
+
+    #if os(iOS)
+      cell.optimize()
+    #endif
+
+    guard indexPath.item < spot.component.items.count else { return cell }
+
+    if let composite = cell as? SpotComposable {
+      let spots = spot.spotsCompositeDelegate?.resolve(spotIndex: spot.index, itemIndex: indexPath.item)
+      composite.configure(&spot.component.items[indexPath.item], spots: spots)
+    } else if let cell = cell as? SpotConfigurable {
+      cell.configure(&spot.component.items[indexPath.item])
+
+      if spot.component.items[indexPath.item].size.height == 0.0 {
+        spot.component.items[indexPath.item].size = cell.size
+      }
+
+      spot.configure?(cell)
+    }
+
+    return cell
+  }
+}
